requirements.txt
openai>=1.0.0
chromadb>=0.4.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
sentence-transformers>=2.2.0
torch>=1.9.0
